pyproject.toml
[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.poetry]
name = "slack-reaction-bot"
description = "Parses slack messages into pythonic objects and manipulates those objects"
version = "0.1.0"
authors = ["Zachary Myers <zachmyers3@gmail.com>"]


[tool.poetry.dependencies]
python = ">=3.9,<3.13"
slack-bolt = "^1.18.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.2"
